app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
langchain==0.2.0
langchain-openai==0.1.8
langchain-community==0.2.0
openai==1.30.0
httpx==0.27.0
numpy==1.26.2
python-docx==0.8.11
pypdf==3.17.1
pytesseract==0.3.10
pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1
pgvector==0.2.4